+++ src/components/PatientDashboard/patient.types.ts
@@ -1,0 +1,12 @@
+export type Patient = {
+  id: number;
+  name: string;
+  middleName: string;
+  lastName: string;
+  age: number;
+  phone: string;
+  mail: string;
+  address: string;
+};
+
+export type PatientList = Patient[];
